refactor: 설정 파일 패키지 분리 및 이름 간략히 변경
