+++ src/main/java/roomescape/common/config/DateTimeFormatConfig.java
@@ -1,4 +1,4 @@
-package roomescape.common;
+package roomescape.common.config;
 
 import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateSerializer;
 import com.fasterxml.jackson.datatype.jsr310.ser.LocalTimeSerializer;
@@ -9,7 +9,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class DateTimeFormatConfiguration {
+public class DateTimeFormatConfig {
     private static final String DATE_FORMAT = "yyyy-MM-dd";
     private static final String TIME_FORMAT = "HH:mm";
