+++ src/main/java/roomescape/common/config/PaymentClientConfig.java
@@ -1,4 +1,4 @@
-package roomescape.common;
+package roomescape.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -6,6 +6,7 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.client.ClientHttpRequestFactory;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.web.client.RestClient;
+import roomescape.common.TossPaymentResponseErrorHandler;
 import roomescape.payment.client.PaymentClient;
 import roomescape.payment.client.PaymentProperties;
 import roomescape.payment.client.TossPaymentClient;
@@ -14,14 +15,14 @@ import java.nio.charset.StandardCharsets;
 import java.util.Base64;
 
 @Configuration
-public class PaymentClientConfiguration {
+public class PaymentClientConfig {
 
     private static final String BASIC_PREFIX = "Basic ";
     private static final String NO_PASSWORD_SUFFIX = ":";
 
     private final PaymentProperties paymentProperties;
 
-    public PaymentClientConfiguration(final PaymentProperties paymentProperties) {
+    public PaymentClientConfig(final PaymentProperties paymentProperties) {
         this.paymentProperties = paymentProperties;
     }
